Add paste RPC and multi-selection paste
This adds a new RPC, 'paste', and modifies the behaviour of paste in the
multi-selection case. Now, if the number of lines to be pasted matches
the number of selection regions, lines will be spread across cursors. If
not, the behaviour is the same.

This also fixes a long standing issue where paste was treated the same
as keyboard input for the purposes of undo. Now, a paste action will
always create its own undo group.

Progress on #323